Addressed comments, added abstract test class
