+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraTimestampBackupRunner.java
@@ -89,7 +89,8 @@ public class CassandraTimestampBackupRunner {
                             Pair.create(currentBackupBound, backupValue));
             executeAndVerifyCas(client, casMap);
             log.info("[BACKUP] Backed up the value {}", currentBound);
-            return CassandraTimestampBoundStoreEntry.createFromBytes(backupValue).getTimestampOrInitialValue();
+            return CassandraTimestampBoundStoreEntry.createFromBytes(backupValue)
+                    .getTimestampOrValue(CassandraTimestampUtils.INITIAL_VALUE);
         });
     }
